class structure refactored
