+++ src/main/java/cat/nyaa/nyaacore/cmdreceiver/NoPermissionException.java
@@ -1,0 +1,10 @@
+package cat.nyaa.nyaacore.cmdreceiver;
+
+class NoPermissionException extends RuntimeException {
+    /**
+     * @param permission name for the permission node
+     */
+    public NoPermissionException(String permission) {
+        super(permission);
+    }
+}
