+++ src/main/java/cat/nyaa/nyaacore/cmdreceiver/NotPlayerException.java
@@ -1,0 +1,5 @@
+package cat.nyaa.nyaacore.cmdreceiver;
+
+class NotPlayerException extends RuntimeException {
+
+}
