+++ src/test/java/cat/nyaa/nyaacore/ArgumentsTest.java
@@ -1,5 +1,7 @@
 package cat.nyaa.nyaacore;
 
+import cat.nyaa.nyaacore.cmdreceiver.Arguments;
+import cat.nyaa.nyaacore.cmdreceiver.BadCommandException;
 import org.junit.Test;
 
 import static org.junit.Assert.*;
@@ -8,7 +10,7 @@ public class ArgumentsTest {
     @Test
     public void test1() throws Exception {
         String cmd = "`foo bar` far `bar \\`foo`";
-        CommandReceiver.Arguments arg = CommandReceiver.Arguments.parse(cmd.split(" "));
+        Arguments arg = Arguments.parse(cmd.split(" "));
         assertNotNull(arg);
         assertEquals("foo bar", arg.nextString());
         assertEquals("far", arg.nextString());
@@ -18,7 +20,7 @@ public class ArgumentsTest {
     @Test
     public void test2() throws Exception {
         String cmd = "key:33 a key:66 b key:13 `key2:a b c` `key 3:d e f`";
-        CommandReceiver.Arguments arg = CommandReceiver.Arguments.parse(cmd.split(" "), null);
+        Arguments arg = Arguments.parse(cmd.split(" "), null);
         assertNotNull(arg);
         assertEquals("key:33", arg.nextString());
         assertEquals(66, arg.argInt("key"));
@@ -34,7 +36,7 @@ public class ArgumentsTest {
     @Test
     public void test3() throws Exception {
         String cmd = "t w key:`3` key2:`/co l u:miu_bug` ke3y:`12`";
-        CommandReceiver.Arguments arg = CommandReceiver.Arguments.parse(cmd.split(" "));
+        Arguments arg = Arguments.parse(cmd.split(" "));
         assertNotNull(arg);
         assertEquals(3, arg.argInt("key"));
         assertEquals("/co l u:miu_bug", arg.argString("key2"));
@@ -43,7 +45,7 @@ public class ArgumentsTest {
         assertEquals("w", arg.nextString());
         try {
             arg.nextString();
-        } catch (CommandReceiver.BadCommandException e) {//seems that no expectThrows available here
+        } catch (BadCommandException e) {//seems that no expectThrows available here
             assertEquals("internal.error.no_more_string", e.getMessage());//magic string, but no impact...
         }
     }
@@ -51,14 +53,14 @@ public class ArgumentsTest {
     @Test
     public void test4() throws Exception {
         String cmd = "key :`3`";
-        CommandReceiver.Arguments arg = CommandReceiver.Arguments.parse(cmd.split(" "), null);
+        Arguments arg = Arguments.parse(cmd.split(" "), null);
         assertNull(arg);
     }
 
     @Test
     public void test5() throws Exception {
         String cmd = "key: `3`";
-        CommandReceiver.Arguments arg = CommandReceiver.Arguments.parse(cmd.split(" "), null);
+        Arguments arg = Arguments.parse(cmd.split(" "), null);
         assertNotNull(arg);
         assertEquals("key:", arg.top());
         assertEquals("", arg.argString("key"));
